+++ dc-model-jackson/src/main/java/de/digitalcollections/model/jackson/mixin/identifiable/entity/work/ManifestationMixIn.java
@@ -3,7 +3,7 @@ package de.digitalcollections.model.jackson.mixin.identifiable.entity.work;
 import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
 import com.fasterxml.jackson.annotation.JsonTypeName;
 import com.fasterxml.jackson.databind.annotation.JsonDeserialize;
-import de.digitalcollections.model.identifiable.entity.work.Manifestation;
+import de.digitalcollections.model.identifiable.entity.manifestation.Manifestation;
 import de.digitalcollections.model.jackson.mixin.identifiable.entity.EntityMixIn;
 
 @JsonDeserialize(as = Manifestation.class)