+++ dc-model-jackson/src/test/java/de/digitalcollections/model/jackson/identifiable/entity/work/ItemTest.java
@@ -1,7 +1,7 @@
 package de.digitalcollections.model.jackson.identifiable.entity.work;
 
 import de.digitalcollections.model.identifiable.Identifier;
-import de.digitalcollections.model.identifiable.entity.work.Item;
+import de.digitalcollections.model.identifiable.entity.item.Item;
 import de.digitalcollections.model.identifiable.resource.ImageFileResource;
 import de.digitalcollections.model.jackson.BaseJsonSerializationTest;
 import de.digitalcollections.model.text.LocalizedText;